Switched to c++ attributes for gnu packed. Rearranged files.
the __attribute__((packed)) has been replaced with [[gnu::packed]] which is neater.
Move files around, moving library modules for debug, c and c++ to a lib folder, extracting the documentation to its own folder, and moving hardware into the boot folder specifically to scope it to this boot and not to the kernel to abstract this.